+++ typescript/src/api/index.ts
@@ -1,7 +1,6 @@
 export {
   type TokensCreateAuthor,
   type TokensCreateInput,
-  type TokensCreateLegacyInput,
   type TokensCreatePrivateKeyInput,
   type TokensCreateResponse,
 } from './resources.js';